(PC-23592)[PRO] feat: Rename SelectAutocomplete2 -> SelectAutocomplete foc component, imports and folder.
